requirements.txt
requests>=2.31.0
websocket-client>=1.6.0
openai>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b0

